+++ src/routes/+layout.svelte
@@ -1,12 +1,7 @@
 <script lang="ts">
   import '../app.css'
-  import favicon from '$lib/assets/favicon.svg'
   import type { LayoutData } from './$types'
   import { injectSpeedInsights } from '@vercel/speed-insights/sveltekit'
-  import { addCollection } from '@iconify/svelte'
-  import arcticons from '@iconify-json/arcticons/icons.json'
-
-  addCollection(arcticons as Parameters<typeof addCollection>[0])
 
   injectSpeedInsights()
 
@@ -17,7 +12,7 @@
 </script>
 
 <svelte:head>
-  <link rel="icon" href={favicon} />
+  <meta name="theme-color" content="#20bbcc" />
 </svelte:head>
 
 {@render children()}